requirements.txt
asyncpg
alembic
pydantic
sqlalchemy
fastapi[all]
python-dotenv
fastapi-users[sqlalchemy]
fastapi-cache2[redis]
aioredis
celery
flower